Tanks on Screen
Added alternative Textures which play nice with custom Colors.

Added a DynamicGameObject class that represents a class that collides with stuff just liked RoboCat did. Since we have Tanks and Projectiles which can collide we cannot take the same approach.

Added a MultiplayerGameState which now makes the necessary calls to establish a connection + use the Renderer etc.

GameObject now has Bounds instead of a Radius, this forces us to give a size to the GameObject though which we specify in the constructor of our inheritors. (It's the image size unscaled!).
GameObject now also has a CanCollide bool.

Changed HandleCollisionWithCat to HandleCollisionWithDynamicGameObject. We now need to make checks with what we collided exactly -> we can use the 4CC to our advantage here and use a switch case.

Added a projectile based on the existing yarn. Untested for now.

Moved WORLD_HEIGHT and WORLD_WIDTH to SharedPCH, everyone has access to it now. It's currently only representative of the screen size, we can introduce more consts later to define the world properly.

Collisions are now handled using bounds.

Added Directions to RoboMath but they probably wont be used as I update rotation based on velocity.

Added Tank based on RoboCat, TankClient based on RoboCatClient and TankServer based on RoboCatServer and did the same thing with Projectile.

PlayerSpriteComponent is now no longer Dependant on RoboCat but uses a GameObject pointer instead.

Server now no longer Spawns Mice, Cats changed to Tank. Use 4CC where applicable to circumvent the removed "GetAsCat" method.

Time Worked: Around 5 Hours.