Pickups + Health Fix
The Server now spawns pickups. I ported some existing stuff from the old codebase but had to make considerable changes to make this work.

The PickupSpawner no longer exists, the PickupSpawnerManager manages all of this himself.

Added the upgrade functionality to the Server Tank. The client does not even know how to apply upgrades which is a plus for security. But that also means that the Missile Upgrade is dropped for now.

Health replication for Tanks now uses 9 bits instead of 4.

Time worked: Around 1h 30.